refactor(settings): modularize api-keys components and achieve 100% react-doctor score
